clould-client/requirements.txt
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
dotenv==0.9.9
filelock==3.18.0
fsspec==2025.3.2
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
openai-whisper==20240930
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydub==0.25.1
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
six==1.17.0
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
